rhymes: Version 4.
Factor out the enumeration of pages within a category to deepCatFilter. I considered using actual deep filtering in precomputeRhymes, but the structure of Wiktionary's English morpheme categories makes this unreasonable.
Call the cached list of words in the appropriate categories the "word cache" rather than the "category cache". Also save the one final list of categorized words rather than the initial lists which are combined to produce the final one.
Include primary stress indicators in pronLines, and use these to exclude words stressed on the wrong syllable.