deepCatFilter: Allow output of ids.
In deepCatFilter, add '-u', which specifies that the output should be given as a list of IDs rather than terms. When the user gives the sets of categories to include and exclude as names rather than IDs and translation is performed, tell them what the IDs are so they can use them directly next time. Improve format of '-v' output.