deepCatFilter: Add form-of filtering.
On Wiktionary (at least the English Wiktionary), forms (like
alternative forms and inflections) are not fully categorized. This
caused terms like "pks" (plural of "pk", an abbreviation of "pack"),
to be included in a search that excluded English abbreviations,
because although it was categorized as a non-lemma, it was not
categorized as an abbreviation. To solve this, search through the
sense lines of each candidate term for form-of templates (descendants
of Category:Form-of templates), and remove the word if every one of
its senses includes a form-of template linking to an excluded term.