find_terms: Add --regex to filter terms by arbitrary regex.
Previously --only-words could be used to specify that only single-word terms
should be included. Upgrade and rename this option to --regex, which now
specifies an arbitrary regex for re.fullmatch() to apply to each term,
rejecting terms that do not match. The old behaviour of selecting single-word
terms can be achieved using:
--regex \w+